+++ postgrator.js
@@ -2,7 +2,6 @@ const fs = require("fs");
 const path = require("path");
 const glob = require("glob");
 const EventEmitter = require("events");
-const deprecate = require("depd")("postgrator");
 
 const createCommonClient = require("./lib/createCommonClient.js");
 const {
@@ -41,39 +40,6 @@ class Postgrator extends EventEmitter {
     this.config = Object.assign({}, DEFAULT_CONFIG, config);
     this.migrations = [];
     this.commonClient = createCommonClient(this.config);
-
-    // Instantiation with database credentials is deprecated
-    // Next major version of postgrator will require user manage the connection
-    // and provide the `execQuery` function
-    if (this.config.port) {
-      deprecate(`Config option "port". Implement execQuery function instead.`);
-    }
-    if (this.config.host) {
-      deprecate(`Config option "host". Implement execQuery function instead.`);
-    }
-    if (this.config.username) {
-      deprecate(
-        `Config option "username". Implement execQuery function instead.`
-      );
-    }
-    if (this.config.password) {
-      deprecate(
-        `Config option "password". Implement execQuery function instead.`
-      );
-    }
-    if (this.config.ssl) {
-      deprecate(`Config option "ssl". Implement execQuery function instead.`);
-    }
-    if (this.config.options) {
-      deprecate(
-        `Config option "options". Implement execQuery function instead.`
-      );
-    }
-    if (this.config.migrationDirectory) {
-      deprecate(
-        `Config option "migrationDirectory". use "migrationPattern" instead using glob match. e.g. path.join(__dirname, '/migrations/*')`
-      );
-    }
   }
 
   /**